Refactored tests and assertions implementation to internal/
