+++ include/criterion/internal/assert.h
@@ -1,31 +1,10 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright © 2015 Franklin "Snaipe" Mathieu <http://snai.pe/>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-#ifndef CRITERION_ASSERT_BASE_H_
-# define CRITERION_ASSERT_BASE_H_
+#ifndef CRITERION_INTERNAL_ASSERT_H_
+# define CRITERION_INTERNAL_ASSERT_H_
 
-# include "internal/preprocess.h"
-# include "internal/asprintf-compat.h"
+# include "common.h"
+# include "preprocess.h"
+# include "asprintf-compat.h"
+# include "designated-initializer-compat.h"
 
 # ifdef __cplusplus
 #  include <cstring>
@@ -35,12 +14,11 @@
 #  include <stdlib.h>
 #  include <stdbool.h>
 # endif
-# include "internal/designated-initializer-compat.h"
-# include "types.h"
-# include "stats.h"
-# include "hooks.h"
-# include "event.h"
-# include "abort.h"
+# include "../types.h"
+# include "../stats.h"
+# include "../hooks.h"
+# include "../event.h"
+# include "../abort.h"
 
 struct criterion_assert_args {
     const char *msg;
@@ -68,12 +46,6 @@ enum criterion_assert_messages {
     CRITERION_ASSERT_MSG_ANY_THROW,
     CRITERION_ASSERT_MSG_NONE_THROW,
 };
-
-CR_BEGIN_C_API
-
-CR_API char *cr_translate_assert_msg(int msg_index, ...);
-
-CR_END_C_API
 
 # define CR_GET_CONDITION(Condition, ...) Condition
 # define CR_GET_CONDITION_STR(Condition, ...) #Condition
@@ -151,8 +123,6 @@ CR_END_C_API
             Fail();                                                         \
     } while (0)
 
-// Base assertions
-
 # define cr_fail(Fail, ...)                                     \
     CR_EXPAND(cr_assert_impl(                                   \
             Fail,                                               \
@@ -163,10 +133,7 @@ CR_END_C_API
             __VA_ARGS__                                         \
     ))
 
-# define cr_assert_fail(...) CR_EXPAND(cr_fail(CR_FAIL_ABORT_,     __VA_ARGS__))
-# define cr_expect_fail(...) CR_EXPAND(cr_fail(CR_FAIL_CONTINUES_, __VA_ARGS__))
-
-# define cr_assert(...)                                                     \
+# define cr_assert_(...)                                                    \
     CR_EXPAND(cr_assert_impl(                                               \
             CR_FAIL_ABORT_,                                                 \
             CR_VA_HEAD(__VA_ARGS__),                                        \
@@ -176,7 +143,7 @@ CR_END_C_API
             CR_VA_TAIL(__VA_ARGS__)                                         \
     ))
 
-# define cr_expect(...)                                                     \
+# define cr_expect_(...)                                                    \
     CR_EXPAND(cr_assert_impl(                                               \
             CR_FAIL_CONTINUES_,                                             \
             CR_VA_HEAD(__VA_ARGS__),                                        \
@@ -186,27 +153,28 @@ CR_END_C_API
             CR_VA_TAIL(__VA_ARGS__)                                         \
     ))
 
-# define cr_assert_not(...)                                                    \
-    CR_EXPAND(cr_assert_impl(                                                  \
-            CR_FAIL_ABORT_,                                                    \
-            !(CR_VA_HEAD(__VA_ARGS__)),                                        \
-            dummy,                                                             \
-            CRITERION_ASSERT_MSG_EXPR_FALSE,                                   \
-            (CR_STR(!(CR_VA_HEAD(__VA_ARGS__)))),                              \
-            CR_VA_TAIL(__VA_ARGS__)                                            \
+# define cr_assert_not_(...)                                                \
+    CR_EXPAND(cr_assert_impl(                                               \
+            CR_FAIL_ABORT_,                                                 \
+            !(CR_VA_HEAD(__VA_ARGS__)),                                     \
+            dummy,                                                          \
+            CRITERION_ASSERT_MSG_EXPR_FALSE,                                \
+            (CR_STR(!(CR_VA_HEAD(__VA_ARGS__)))),                           \
+            CR_VA_TAIL(__VA_ARGS__)                                         \
     ))
 
-# define cr_expect_not(...)                                                    \
-    CR_EXPAND(cr_assert_impl(                                                  \
-            CR_FAIL_CONTINUES_,                                                \
-            !(CR_VA_HEAD(__VA_ARGS__)),                                        \
-            dummy,                                                             \
-            CRITERION_ASSERT_MSG_EXPR_FALSE,                                   \
-            (CR_STR(!(CR_VA_HEAD(__VA_ARGS__)))),                              \
-            CR_VA_TAIL(__VA_ARGS__)                                            \
+# define cr_expect_not_(...)                                                 \
+    CR_EXPAND(cr_assert_impl(                                               \
+            CR_FAIL_CONTINUES_,                                             \
+            !(CR_VA_HEAD(__VA_ARGS__)),                                     \
+            dummy,                                                          \
+            CRITERION_ASSERT_MSG_EXPR_FALSE,                                \
+            (CR_STR(!(CR_VA_HEAD(__VA_ARGS__)))),                           \
+            CR_VA_TAIL(__VA_ARGS__)                                         \
     ))
 
-// Common binary assertions
+
+// Binary
 
 # define cr_assert_op_(Fail, Op, Actual, Expected, ...)                     \
     CR_EXPAND(cr_assert_impl(                                               \
@@ -227,25 +195,7 @@ CR_END_C_API
             CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__))                             \
     ))
 
-# define cr_assert_eq(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_ABORT_,      ==, __VA_ARGS__))
-# define cr_expect_eq(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_CONTINUES_,  ==, __VA_ARGS__))
-
-# define cr_assert_neq(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_ABORT_,     !=, __VA_ARGS__))
-# define cr_expect_neq(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_CONTINUES_, !=, __VA_ARGS__))
-
-# define cr_assert_lt(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_ABORT_,      <, __VA_ARGS__))
-# define cr_expect_lt(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_CONTINUES_,  <, __VA_ARGS__))
-
-# define cr_assert_leq(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_ABORT_,     <=, __VA_ARGS__))
-# define cr_expect_leq(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_CONTINUES_, <=, __VA_ARGS__))
-
-# define cr_assert_gt(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_ABORT_,      >, __VA_ARGS__))
-# define cr_expect_gt(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_CONTINUES_,  >, __VA_ARGS__))
-
-# define cr_assert_geq(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_ABORT_,     >=, __VA_ARGS__))
-# define cr_expect_geq(...) CR_EXPAND(cr_assert_op_va_(CR_FAIL_CONTINUES_, >=, __VA_ARGS__))
-
-// Common unary assertions
+// Unary
 
 # define cr_assert_null_op_(Fail, Op, Msg, Value, ...)  \
     CR_EXPAND(cr_assert_impl(                           \
@@ -266,13 +216,7 @@ CR_END_C_API
             CR_VA_TAIL(__VA_ARGS__)                     \
     ))
 
-# define cr_assert_null(...) CR_EXPAND(cr_assert_null_op_va_(CR_FAIL_ABORT_,     ==, CRITERION_ASSERT_MSG_IS_NOT_NULL, __VA_ARGS__))
-# define cr_expect_null(...) CR_EXPAND(cr_assert_null_op_va_(CR_FAIL_CONTINUES_, ==, CRITERION_ASSERT_MSG_IS_NOT_NULL, __VA_ARGS__))
-
-# define cr_assert_not_null(...) CR_EXPAND(cr_assert_null_op_va_(CR_FAIL_ABORT_,     !=, CRITERION_ASSERT_MSG_IS_NULL, __VA_ARGS__))
-# define cr_expect_not_null(...) CR_EXPAND(cr_assert_null_op_va_(CR_FAIL_CONTINUES_, !=, CRITERION_ASSERT_MSG_IS_NULL, __VA_ARGS__))
-
-// Floating-point assertions
+// Floating point
 
 # define cr_assert_float_eq_op_(Actual, Expected, Epsilon) \
     (Expected) - (Actual) <= (Epsilon) && (Actual) - (Expected) <= (Epsilon)
@@ -300,13 +244,7 @@ CR_END_C_API
             CR_VA_TAIL(CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__)))     \
     ))
 
-# define cr_assert_float_eq(...) CR_EXPAND(cr_assert_float_op_va_(CR_FAIL_ABORT_,     cr_assert_float_eq_op_, __VA_ARGS__))
-# define cr_expect_float_eq(...) CR_EXPAND(cr_assert_float_op_va_(CR_FAIL_CONTINUES_, cr_assert_float_eq_op_, __VA_ARGS__))
-
-# define cr_assert_float_neq(...) CR_EXPAND(cr_assert_float_op_va_(CR_FAIL_ABORT_,     cr_assert_float_neq_op_, __VA_ARGS__))
-# define cr_expect_float_neq(...) CR_EXPAND(cr_assert_float_op_va_(CR_FAIL_CONTINUES_, cr_assert_float_neq_op_, __VA_ARGS__))
-
-// String assertions
+// String
 
 # define cr_assert_str_op_empty_(Fail, Op, Msg, Value, ...) \
     CR_EXPAND(cr_assert_impl(                               \
@@ -327,12 +265,6 @@ CR_END_C_API
             CR_VA_TAIL(__VA_ARGS__)                     \
     ))
 
-# define cr_assert_str_empty(...) CR_EXPAND(cr_assert_str_op_empty_va_(CR_FAIL_ABORT_, ==, CRITERION_ASSERT_MSG_IS_NOT_EMPTY, __VA_ARGS__))
-# define cr_expect_str_empty(...) CR_EXPAND(cr_assert_str_op_empty_va_(CR_FAIL_CONTINUES_, ==, CRITERION_ASSERT_MSG_IS_NOT_EMPTY, __VA_ARGS__))
-
-# define cr_assert_str_not_empty(...) CR_EXPAND(cr_assert_str_op_empty_va_(CR_FAIL_ABORT_, !=, CRITERION_ASSERT_MSG_IS_EMPTY, __VA_ARGS__))
-# define cr_expect_str_not_empty(...) CR_EXPAND(cr_assert_str_op_empty_va_(CR_FAIL_CONTINUES_, !=, CRITERION_ASSERT_MSG_IS_EMPTY, __VA_ARGS__))
-
 # define cr_assert_str_op_(Fail, Op, Actual, Expected, ...) \
     CR_EXPAND(cr_assert_impl(                               \
             Fail,                                           \
@@ -352,22 +284,165 @@ CR_END_C_API
             CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__))     \
     ))
 
-# define cr_assert_str_eq(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_ABORT_,      ==, __VA_ARGS__))
-# define cr_expect_str_eq(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_CONTINUES_,  ==, __VA_ARGS__))
+// Array
 
-# define cr_assert_str_neq(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_ABORT_,     !=, __VA_ARGS__))
-# define cr_expect_str_neq(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_CONTINUES_, !=, __VA_ARGS__))
+# define cr_assert_mem_op_(Fail, Op, Actual, Expected, Size, ...)       \
+    CR_EXPAND(cr_assert_impl(                                           \
+            Fail,                                                       \
+            CR_STDN memcmp((Actual), (Expected), (Size)) Op 0,          \
+            dummy,                                                      \
+            CRITERION_ASSERT_MSG_EXPR_FALSE,                            \
+            (CR_STR((Actual)[0 .. Size] Op (Expected)[0 .. Size])),     \
+            __VA_ARGS__                                                 \
+    ))
 
-# define cr_assert_str_lt(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_ABORT_,      <, __VA_ARGS__))
-# define cr_expect_str_lt(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_CONTINUES_,  <, __VA_ARGS__))
+# define cr_assert_mem_op_va_(Fail, Op, ...)                    \
+    CR_EXPAND(cr_assert_mem_op_(                                \
+            Fail,                                               \
+            Op,                                                 \
+            CR_VA_HEAD(__VA_ARGS__),                            \
+            CR_VA_HEAD(CR_VA_TAIL(__VA_ARGS__)),                \
+            CR_VA_HEAD(CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__))),    \
+            CR_VA_TAIL(CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__)))     \
+    ))
 
-# define cr_assert_str_leq(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_ABORT_,     <=, __VA_ARGS__))
-# define cr_expect_str_leq(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_CONTINUES_, <=, __VA_ARGS__))
+// Array comparisons
 
-# define cr_assert_str_gt(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_ABORT_,      >, __VA_ARGS__))
-# define cr_expect_str_gt(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_CONTINUES_,  >, __VA_ARGS__))
+#  ifdef __cplusplus
+#   define CR_ARR_COMPARE_(A, B, Size, Cmp, Result)                     \
+    int Result = std::lexicographical_compare((A), (A) + Size, (B), (B) + Size, Cmp)
+#  else
+#   define CR_ARR_COMPARE_(A, B, Size, Cmp, Result)                     \
+        __typeof__(&(A)[0]) first = (A);                                \
+        __typeof__(&(B)[0]) second = (B);                               \
+        int Result = 0;                                                 \
+        size_t i, size;                                                 \
+        for (i = 0, size = (Size); !Result && i < size; ++i)            \
+            Result = Cmp(first + i, second + i)
+#  endif
 
-# define cr_assert_str_geq(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_ABORT_,     >=, __VA_ARGS__))
-# define cr_expect_str_geq(...) CR_EXPAND(cr_assert_str_op_va_(CR_FAIL_CONTINUES_, >=, __VA_ARGS__))
+#  define cr_assert_arr_op_cmp_(Fail, Op, Actual, Expected, Size, Cmp, ...) \
+    do {                                                                    \
+        CR_ARR_COMPARE_(Actual, Expected, Size, Cmp, order);                \
+        CR_EXPAND(cr_assert_impl(                                           \
+                Fail,                                                       \
+                order Op 0,                                                 \
+                dummy,                                                      \
+                CRITERION_ASSERT_MSG_EXPR_FALSE,                            \
+                (CR_STR((Actual)[0 .. Size] Op (Expected)[0 .. Size])),     \
+                __VA_ARGS__                                                 \
+        ));                                                                 \
+    } while (0)
 
-#endif /* !CRITERION_ASSERT_BASE_H_ */
+#  define cr_assert_arr_op_cmp_va_(Fail, Op, ...)                           \
+    CR_EXPAND(cr_assert_arr_op_cmp_(                                        \
+            Fail,                                                           \
+            Op,                                                             \
+            CR_VA_HEAD(__VA_ARGS__),                                        \
+            CR_VA_HEAD(CR_VA_TAIL(__VA_ARGS__)),                            \
+            CR_VA_HEAD(CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__))),                \
+            CR_VA_HEAD(CR_VA_TAIL(CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__)))),    \
+            CR_VA_TAIL(CR_VA_TAIL(CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__))))     \
+    ))
+
+// Exceptions
+
+#  define cr_assert_throw_abort_(Fail, Msg, MsgArgs, ...)               \
+    CR_EXPAND(cr_assert_impl(                                           \
+            Fail,                                                       \
+            0,                                                          \
+            dummy,                                                      \
+            Msg,                                                        \
+            MsgArgs,                                                    \
+            CR_VA_TAIL(__VA_ARGS__)                                     \
+    ))
+
+#  define cr_assert_throw_(Fail, Statement, Exception, ...)             \
+    try {                                                               \
+        Statement;                                                      \
+    } catch (Exception const &) {                                       \
+    } catch (...) {                                                     \
+        CR_EXPAND(cr_assert_throw_abort_(                               \
+                Fail,                                                   \
+                CRITERION_ASSERT_MSG_NO_THROW,                          \
+                (CR_STR(Statement), CR_STR(Exception)),                 \
+                __VA_ARGS__));                                          \
+    }
+
+#  define cr_assert_throw_va_(...)                                      \
+    CR_EXPAND(cr_assert_throw_(                                         \
+            CR_VA_HEAD(__VA_ARGS__),                                    \
+            CR_VA_HEAD(CR_VA_TAIL(__VA_ARGS__)),                        \
+            CR_VA_HEAD(CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__))),            \
+            dummy,                                                      \
+            CR_VA_TAIL(CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__)))             \
+    ))
+
+#  define cr_assert_no_throw_(Fail, Statement, Exception, ...)          \
+    try {                                                               \
+        Statement;                                                      \
+    } catch (Exception const &) {                                       \
+        CR_EXPAND(cr_assert_throw_abort_(                               \
+                Fail,                                                   \
+                CRITERION_ASSERT_MSG_THROW,                             \
+                (CR_STR(Statement), CR_STR(Exception)),                 \
+                __VA_ARGS__));                                          \
+    }
+
+#  define cr_assert_no_throw_va_(...)                                   \
+    CR_EXPAND(cr_assert_no_throw_(                                      \
+            CR_VA_HEAD(__VA_ARGS__),                                    \
+            CR_VA_HEAD(CR_VA_TAIL(__VA_ARGS__)),                        \
+            CR_VA_HEAD(CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__))),            \
+            dummy,                                                      \
+            CR_VA_TAIL(CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__)))             \
+    ))
+
+#  define cr_assert_any_throw_(Fail, Statement, ...)                    \
+    try {                                                               \
+        Statement;                                                      \
+        CR_EXPAND(cr_assert_throw_abort_(                               \
+                Fail,                                                   \
+                CRITERION_ASSERT_MSG_ANY_THROW,                         \
+                (CR_STR(Statement)),                                    \
+                __VA_ARGS__));                                          \
+    } catch (...) {}
+
+#  define cr_assert_any_throw_va_(...)                                  \
+    CR_EXPAND(cr_assert_any_throw_(                                     \
+            CR_VA_HEAD(__VA_ARGS__),                                    \
+            CR_VA_HEAD(CR_VA_TAIL(__VA_ARGS__)),                        \
+            dummy,                                                      \
+            CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__))                         \
+    ))
+
+#  define cr_assert_none_throw_(Fail, Statement, ...)                   \
+    try {                                                               \
+        Statement;                                                      \
+    } catch (...) {                                                     \
+        CR_EXPAND(cr_assert_throw_abort_(                               \
+                Fail,                                                   \
+                CRITERION_ASSERT_MSG_NONE_THROW,                        \
+                (CR_STR(Statement)),                                    \
+                __VA_ARGS__));                                          \
+    }
+
+#  define cr_assert_none_throw_va_(...)                                 \
+    CR_EXPAND(cr_assert_none_throw_(                                    \
+            CR_VA_HEAD(__VA_ARGS__),                                    \
+            CR_VA_HEAD(CR_VA_TAIL(__VA_ARGS__)),                        \
+            dummy,                                                      \
+            CR_VA_TAIL(CR_VA_TAIL(__VA_ARGS__))                         \
+    ))
+
+// Messages
+
+#  define CRITERION_GNUC_WARN__(Msg) \
+    _Pragma(#Msg)
+
+#  define CRITERION_GNUC_WARN_(Name) CRITERION_GNUC_WARN__(             \
+        message                                                         \
+        "The `" #Name "` macro is only available on GNU C compilers."   \
+    )
+
+#endif /* !CRITERION_INTERNAL_ASSERT_H_ */
