+++ include/criterion/theories.h
@@ -122,7 +122,7 @@ CR_API void cr_theory_main(struct criterion_datapoints *dps, size_t datapoints, 
 
 # define Theory(Args, ...)                                                      \
     void CR_EXPAND(CR_VAARG_ID(theory, __VA_ARGS__,))Args;                      \
-    CR_EXPAND(Test_(__VA_ARGS__, .sentinel_ = 0)) {                             \
+    CR_EXPAND(CR_TEST_BASE(__VA_ARGS__, .sentinel_ = 0)) {                      \
         cr_theory_main(                                                         \
                 CR_EXPAND(CR_VAARG_ID(dps, __VA_ARGS__,)),                      \
                 CR_NB_DATAPOINTS(CR_EXPAND(CR_VAARG_ID(dps, __VA_ARGS__,))),    \
